requirements.txt
awxkit>=13.0.0
boto>=2.49.0
botocore>=1.12.249
boto3>=1.9.249
openshift>=0.6.2
requests-oauthlib
durable-rules
pyparsing==3.0.7
azure-servicebus==7.5.0
azure-common==1.1.27
azure-core==1.21.1
ansible-events==0.1.2
watchdog==0.9.0
dpath==2.0.6
Flask==2.0.3
gevent==21.12.0
redis==4.1.4
